requirements.txt
alembic==1.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
Flask-Migrate==3.1.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
pipreqs==0.4.11
psycopg==3.1.3
psycopg2==2.9.4
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
